feat; added PublicSquadSubmissionActions.tsx and used for common behaviour;
